update some server logic

+++ web/server.go
@@ -1,14 +1,19 @@
 package web
 
 import (
+	"context"
 	"fmt"
+	"io/ioutil"
 	"net/http"
 	"os"
 	"os/signal"
+	"strconv"
 	"strings"
 	"syscall"
+	"time"
 
 	"github.com/gookit/event"
+	"github.com/gookit/goutil/sysutil"
 	"github.com/gookit/lako"
 )
 
@@ -47,42 +52,100 @@ func (s *HTTPServer) Start() error {
 
 	s.srv = &http.Server{
 		Addr: s.realAddr,
-		Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			defer func() {
-				if err := recover(); err != nil {
-					w.WriteHeader(500)
-				}
-			}()
-
-			evtData := event.M{"w": w, "r": r}
-
-			// Fire before route
-			app.MustFire(OnBeforeRoute, evtData)
-
-			// Route and dispatch request
-			app.Router.ServeHTTP(w, r)
-
-			// Fire after route
-			app.MustFire(OnAfterRoute, evtData)
-		}),
 	}
+
+	s.srv.Handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		defer func() {
+			if err := recover(); err != nil {
+				w.WriteHeader(500)
+			}
+		}()
+
+		evtData := event.M{"w": w, "r": r}
+
+		// Fire before route
+		app.MustFire(OnBeforeRoute, evtData)
+
+		// Route and dispatch request
+		app.Router.ServeHTTP(w, r)
+
+		// Fire after route
+		app.MustFire(OnAfterRoute, evtData)
+	})
 
 	app.MustFire(OnServerStart, event.M{"addr": s.srv.Addr})
 
-	// listen signal
+	// Listen signal
 	s.handleSignal(s.srv)
 
-	err := s.srv.ListenAndServe()
+	// Save pid to file
+	savePidToFile(s.processID, s.pidFile)
 
+	// Start server
+	err := s.srv.ListenAndServe()
 	if err != http.ErrServerClosed {
 		return err
 	}
 	return nil
 }
 
+// Stop by send signal to exists process
+func (s *HTTPServer) Stop(timeout int) error {
+	pid := s.processID
+	if pid <= 0 {
+		return fmt.Errorf("invalid process ID value")
+	}
+
+	// Server is not started
+	if !sysutil.ProcessExists(pid) {
+		return fmt.Errorf("cannot stop, the process is not exists(PID: %d)", pid)
+	}
+
+	return syscall.Kill(pid, syscall.SIGTERM)
+}
+
+// Shutdown from internal
+func (s *HTTPServer) Shutdown() error {
+	ctx, cancel := context.WithTimeout(context.TODO(), 5*time.Second)
+
+	defer cancel()
+	return s.srv.Shutdown(ctx)
+}
+
 /*************************************************************
  * Getter/Setter methods
  *************************************************************/
+
+// IsRunning get server is running
+func (s *HTTPServer) IsRunning() bool {
+	if s.srv != nil {
+		return true
+	}
+
+	// Get stat by pid
+	if s.pidFile == "" {
+		return false
+	}
+
+	// Get pid from file
+	bts, err := ioutil.ReadFile(s.pidFile)
+	if err != nil {
+		// panic(err)
+		return false
+	}
+
+	pid, _ := strconv.Atoi(string(bts))
+	if pid <= 0 {
+		// panic("invalid process ID value, read from pidFile")
+		return false
+	}
+
+	// Storage pid value
+	s.processID = pid
+
+	// Server is not started
+	return sysutil.ProcessExists(pid)
+}
 
 // RealAddr get resolved read addr
 func (s *HTTPServer) RealAddr() string {
@@ -112,7 +175,7 @@ func (s *HTTPServer) handleSignal(server *http.Server) {
 	go func() {
 		s := <-c
 		fmt.Printf("Got signal [%s], exiting server now", s)
-		if err := server.Close(); err != nil{
+		if err := server.Close(); err != nil {
 			fmt.Printf("Server close failed: %s", err.Error())
 		}
 
@@ -124,7 +187,23 @@ func (s *HTTPServer) handleSignal(server *http.Server) {
 	}()
 }
 
+func savePidToFile(pid int, pidFile string) {
+	if pidFile == "" {
+		return
+	}
+
+	txt := fmt.Sprintf("%d", pid)
+	err := ioutil.WriteFile(pidFile, []byte(txt), 0664)
+	if err != nil {
+		panic(err)
+	}
+}
+
 func removePidFile(pidFile string) error {
+	if pidFile == "" {
+		return nil
+	}
+
 	return os.Remove(pidFile)
 }
